requirements.txt
pandas==2.2.3
streamlit==1.45.0   
google-generativeai==0.16.0
lxml==5.1.0
plotly==5.20.0